refactor: separate data and domain layers
