+++ app/src/main/kotlin/com/example/leveluplccd/MainActivity.kt
@@ -11,9 +11,10 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import androidx.navigation.compose.composable
 import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
+import com.example.leveluplccd.ui.quest.DailyQuestScreen
 import com.example.leveluplccd.ui.theme.LevelUpLccdTheme
 
 class MainActivity : ComponentActivity() {
@@ -31,7 +32,7 @@ fun LevelUpLccdApp() {
             composable(Destinations.Home.route) {
                 HomeScreen(onNavigate = { navController.navigate(it.route) })
             }
-            composable(Destinations.Quest.route) { com.example.leveluplccd.quest.DailyQuestScreen() }
+            composable(Destinations.Quest.route) { DailyQuestScreen() }
             composable(Destinations.Career.route) { CareerScreen() }
             composable(Destinations.Leaderboard.route) { LeaderboardScreen() }
         }
