+++ app/src/main/kotlin/com/example/leveluplccd/data/QuestRepository.kt
@@ -1,4 +1,4 @@
-package com.example.leveluplccd.quest
+package com.example.leveluplccd.data
 
 import android.content.Context
 import androidx.datastore.core.DataStore
@@ -6,6 +6,8 @@ import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.edit
 import androidx.datastore.preferences.core.intPreferencesKey
 import androidx.datastore.preferences.preferencesDataStore
+import com.example.leveluplccd.domain.Quest
+import com.example.leveluplccd.domain.QuestOption
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.map
