+++ app/src/test/kotlin/com/example/leveluplccd/data/QuestRepositoryTest.kt
@@ -1,4 +1,4 @@
-package com.example.leveluplccd.quest
+package com.example.leveluplccd.data
 
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
